Gbnoisebranch

+++ src/appuser/appuser.controller.ts
@@ -1,15 +1,21 @@
-import { Body, Controller, Delete, Get, HttpStatus, Param, Post, Put, Query, Req, Res } from '@nestjs/common';
+import { Body, Controller, Delete, ForbiddenException, Get, HttpStatus, Param, Post, Put, Query, Req, Request, Res, UseGuards } from '@nestjs/common';
 import { Prisma } from '@prisma/client';
 import { Response } from 'express';
 import { AppUserService } from './appuser.service';
-import { APP_USER_CONTROLLER_ROUTE } from './appuser.utils';
+import { APP_USER_CONTROLLER_ROUTE, validateAdmin, validateSameUser } from './appuser.utils';
+import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { RolesGuard } from 'src/auth/roles.guard';
+import { HasRoles } from 'src/auth/roles.decorator';
+import { Roles } from 'src/auth/auth.utils';
 
 @Controller(APP_USER_CONTROLLER_ROUTE)
 export class AppUserController {
   constructor(private readonly appUserService: AppUserService) { }
 
+  @HasRoles(Roles.ROLE_ADMIN)
+  @UseGuards(JwtAuthGuard, RolesGuard)
   @Get()
-  async getAllUsers(@Res() res: Response, @Query('showRoles') showRoles: boolean, @Query('page') page: number) {
+  async getAllUsers(@Request() req: any, @Res() res: Response, @Query('showRoles') showRoles: boolean, @Query('page') page: number) {
     try {
       const response = await this.appUserService.getAllUsers({ showRoles, page });
       return res.status(200).send(response);
@@ -18,14 +24,17 @@ export class AppUserController {
     }
   }
 
+  @HasRoles(Roles.ROLE_USER)
+  @UseGuards(JwtAuthGuard, RolesGuard)
   @Get('/user')
-  async getSingleUser(@Query('id') id: string, @Query('username') username: string, @Res() res: Response) {
+  async getSingleUser(@Request() req: any, @Query('id') id: string, @Query('username') username: string, @Res() res: Response, @Query('showRoles') showRoles: string) {
     try {
-      console.log({ id, username });
-      const response = await this.appUserService.getUser(id, username);
+      if (!validateSameUser(req.user, id, username)) throw new ForbiddenException()
+
+      const response = await this.appUserService.getUser(id, username, { showRoles });
       return res.status(HttpStatus.OK).send(response);
     } catch (e) {
-      return res.status(HttpStatus.INTERNAL_SERVER_ERROR).send(e);
+      return res.status(e.status).send(e);
     }
   }
 
@@ -39,9 +48,16 @@ export class AppUserController {
     }
   }
 
+  @HasRoles(Roles.ROLE_USER)
+  @UseGuards(JwtAuthGuard, RolesGuard)
   @Put()
-  async updateUser(@Res() res: Response, @Body() user: Prisma.UserUpdateInput) {
+  async updateUser(@Request() req: any, @Res() res: Response, @Body() user: Prisma.UserUpdateInput) {
     try {
+      // once this is done we need to update the access_token because if this endpoint updates the user username
+      // we will no longer be able to update using the username nor doing anything related with the req.user.username
+      // as it will no longer be the same
+      if (!validateSameUser(req.user, user.id as string, user.username as string))
+        throw new ForbiddenException()
       const response = await this.appUserService.updateUser(user);
       return res.status(HttpStatus.OK).send(response);
     } catch (e) {
@@ -49,9 +65,15 @@ export class AppUserController {
     }
   }
 
+
+  @HasRoles(Roles.ROLE_USER)
+  @UseGuards(JwtAuthGuard, RolesGuard)
   @Delete('/:id')
-  async deleteUser(@Res() res: Response, @Param('id') id: string) {
+  async deleteUser(@Request() req: any, @Res() res: Response, @Param('id') id: string) {
     try {
+      if (!validateSameUser(req.user, id, undefined) && !validateAdmin(req.user))
+        throw new ForbiddenException()
+
       const response = await this.appUserService.deleteUser(id);
       return res.status(HttpStatus.OK).send(response);
     } catch (e) {
